update fake account client
